+++ PoshMcp.Server/PowerShell/PowerShellConfiguration.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using PoshMcp.Server.PowerShell.OutOfProcess;
 
 namespace PoshMcp.Server.PowerShell;